refactor: RoomType 폴더명 roomtypes로 변경
