+++ src/main/java/com/beyond/easycheck/roomtypes/infrastructure/repository/RoomTypeRepository.java
@@ -1,6 +1,6 @@
-package com.beyond.easycheck.roomType.infrastructure.repository;
+package com.beyond.easycheck.roomtypes.infrastructure.repository;
 
-import com.beyond.easycheck.roomType.infrastructure.entity.RoomTypeEntity;
+import com.beyond.easycheck.roomtypes.infrastructure.entity.RoomTypeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomTypeRepository extends JpaRepository<RoomTypeEntity, Long> {
